requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
openai==1.55.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
asyncpg==0.30.0
pydantic==2.9.2
alembic==1.17.2
better-auth==1.4.10
psycopg2-binary==2.9.9